move pyscf interop tests
